refactor(router-store): change serializer names
