+++ modules/router-store/src/index.ts
@@ -36,12 +36,12 @@ export {
   BaseRouterStoreState,
 } from './serializers/base';
 export {
-  DefaultRouterStateSerializer,
+  FullRouterStateSerializer,
   SerializedRouterStateSnapshot,
-} from './serializers/default_serializer';
+} from './serializers/full_serializer';
 export {
   MinimalActivatedRouteSnapshot,
   MinimalRouterStateSnapshot,
-  MinimalRouterStateSerializer,
-} from './serializers/minimal_serializer';
+  DefaultRouterStateSerializer,
+} from './serializers/default_serializer';
 export { getSelectors, createRouterSelector } from './router_selectors';
