+++ modules/router-store/src/router_store_module.ts
@@ -39,18 +39,18 @@ import {
   BaseRouterStoreState,
 } from './serializers/base';
 import {
-  DefaultRouterStateSerializer,
+  FullRouterStateSerializer,
   SerializedRouterStateSnapshot,
-} from './serializers/default_serializer';
-import { MinimalRouterStateSerializer } from './serializers/minimal_serializer';
+} from './serializers/full_serializer';
+import { DefaultRouterStateSerializer } from './serializers/default_serializer';
 
 export type StateKeyOrSelector<
   T extends BaseRouterStoreState = SerializedRouterStateSnapshot
 > = string | Selector<any, RouterReducerState<T>>;
 
 /**
- * Full = Serializes the router event with DefaultRouterStateSerializer
- * Minimal = Serializes the router event with MinimalRouterStateSerializer
+ * Full = Serializes the router event with FullRouterStateSerializer
+ * Minimal = Serializes the router event with DefaultRouterStateSerializer
  */
 export const enum RouterState {
   Full,
@@ -72,8 +72,8 @@ export interface StoreRouterConfig<
   navigationActionTiming?: NavigationActionTiming;
   /**
    * Decides which router serializer should be used, if there is none provided, and the metadata on the dispatched @ngrx/router-store action payload.
-   * Set to `Full` to use the `DefaultRouterStateSerializer` and to set the angular router events as payload.
-   * Set to `Minimal` to use the `MinimalRouterStateSerializer` and to set a minimal router event with the navigation id and url as payload.
+   * Set to `Full` to use the `FullRouterStateSerializer` and to set the angular router events as payload.
+   * Set to `Minimal` to use the `DefaultRouterStateSerializer` and to set a minimal router event with the navigation id and url as payload.
    */
   routerState?: RouterState;
 }
@@ -102,7 +102,7 @@ export function _createRouterConfig(
 ): StoreRouterConfig {
   return {
     stateKey: DEFAULT_ROUTER_FEATURENAME,
-    serializer: MinimalRouterStateSerializer,
+    serializer: DefaultRouterStateSerializer,
     navigationActionTiming: NavigationActionTiming.PreActivation,
     ...config,
   };
@@ -183,8 +183,8 @@ export class StoreRouterConnectingModule {
           useClass: config.serializer
             ? config.serializer
             : config.routerState === RouterState.Full
-            ? DefaultRouterStateSerializer
-            : MinimalRouterStateSerializer,
+            ? FullRouterStateSerializer
+            : DefaultRouterStateSerializer,
         },
       ],
     };
@@ -206,14 +206,14 @@ export class StoreRouterConnectingModule {
       isDevMode() &&
       (activeRuntimeChecks?.strictActionSerializability ||
         activeRuntimeChecks?.strictStateSerializability) &&
-      this.serializer instanceof DefaultRouterStateSerializer
+      this.serializer instanceof FullRouterStateSerializer
     ) {
       console.warn(
         '@ngrx/router-store: The serializability runtime checks cannot be enabled ' +
-          'with the DefaultRouterStateSerializer. The default serializer ' +
+          'with the FullRouterStateSerializer. The FullRouterStateSerializer ' +
           'has an unserializable router state and actions that are not serializable. ' +
           'To use the serializability runtime checks either use ' +
-          'the MinimalRouterStateSerializer or implement a custom router state serializer. ' +
+          'the DefaultRouterStateSerializer or implement a custom router state serializer. ' +
           'This also applies to Ivy with immutability runtime checks.'
       );
     }
