+++ modules/router-store/src/serializers/full_serializer.ts
@@ -1,0 +1,54 @@
+import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { BaseRouterStoreState, RouterStateSerializer } from './base';
+
+export interface SerializedRouterStateSnapshot extends BaseRouterStoreState {
+  root: ActivatedRouteSnapshot;
+  url: string;
+}
+
+/**
+ * @deprecated The FullRouterStateSerializer cannot be used when serializability runtime checks are enabled. Use the {@link DefaultRouterStateSerializer} or a {@link https://ngrx.io/guide/router-store/configuration#custom-router-state-serializer custom serializer} instead.
+ */
+export class FullRouterStateSerializer
+  implements RouterStateSerializer<SerializedRouterStateSnapshot>
+{
+  serialize(routerState: RouterStateSnapshot): SerializedRouterStateSnapshot {
+    return {
+      root: this.serializeRoute(routerState.root),
+      url: routerState.url,
+    };
+  }
+
+  private serializeRoute(
+    route: ActivatedRouteSnapshot
+  ): ActivatedRouteSnapshot {
+    const children = route.children.map((c) => this.serializeRoute(c));
+    return {
+      params: route.params,
+      paramMap: route.paramMap,
+      data: route.data,
+      url: route.url,
+      outlet: route.outlet,
+      routeConfig: route.routeConfig
+        ? {
+            component: route.routeConfig.component,
+            path: route.routeConfig.path,
+            pathMatch: route.routeConfig.pathMatch,
+            redirectTo: route.routeConfig.redirectTo,
+            outlet: route.routeConfig.outlet,
+          }
+        : null,
+      queryParams: route.queryParams,
+      queryParamMap: route.queryParamMap,
+      fragment: route.fragment,
+      component: (route.routeConfig
+        ? route.routeConfig.component
+        : undefined) as any,
+      root: undefined as any,
+      parent: undefined as any,
+      firstChild: children[0],
+      pathFromRoot: undefined as any,
+      children,
+    };
+  }
+}
