+++ projects/ngrx.io/content/guide/router-store/configuration.md
@@ -10,20 +10,9 @@ interface StoreRouterConfig {
 </code-example>
 
 - `stateKey`: The name of reducer key, defaults to `router`. It's also possible to provide a selector function.
-- `serializer`: How a router snapshot is serialized. Defaults to `DefaultRouterStateSerializer` (see [Default Router State Serializer](#default-router-state-serializer)). See [Custom Router State Serializer](#custom-router-state-serializer) for more information.
+- `serializer`: How a router snapshot is serialized. Defaults to `DefaultRouterStateSerializer`. See [Custom Router State Serializer](#custom-router-state-serializer) for more information.
 - `navigationActionTiming`: When the `ROUTER_NAVIGATION` is dispatched. Defaults to `NavigationActionTiming.PreActivation`. See [Navigation Action Timing](#navigation-action-timing) for more information.
 - `routerState`: Set this property to decide which serializer should be used, if none is provided, and the metadata on the dispatched action.
-
-## Default Router State Serializer
-
-`DefaultRouterStateSerializer` router state serializer, serializes the URL together with the [ActivatedRouteSnapshot](https://angular.io/api/router/ActivatedRouteSnapshot) from [Angular Router](https://angular.io/guide/router). The latter is serialized recursively, but only with the possibility to traverse the route downward since `root` and `parent` parameters are set to `undefined`.
-
-<div class="alert is-important">
-
-The `DefaultRouterStateSerializer` cannot be used when [serializability runtime checks](guide/store/configuration/runtime-checks) are enabled.
-With serializability runtime checks enabled, the `MinimalRouterStateSerializer` serializer **must** be used. This also applies to Ivy with immutability runtime checks.
-
-</div>
 
 ## Custom Router State Serializer
 
@@ -110,9 +99,21 @@ StoreRouterConnectingModule.forRoot({
 
 This property decides which router serializer should be used. If there is a custom serializer provided, it will use the provided serializer. `routerState` also sets the metadata on dispatched `@ngrx/router-store` action.
 
+### RouterState.Minimal
+
+`RouterState.Minimal` will use the `DefaultRouterStateSerializer` serializer to serialize the Angular Router's `RouterState` and `RouterEvent`.
+
+The difference between `FullRouterStateSerializer` and the `DefaultRouterStateSerializer` is that this serializer is fully serializable. To make the state and the actions serializable, the properties `paramMap`, `queryParamMap` and `component` are ignored.
+
+<code-example header="app.module.ts">
+StoreRouterConnectingModule.forRoot({
+  routerState: RouterState.Minimal,
+});
+</code-example>
+
 ### RouterState.Full
 
-When this property is set to `RouterState.Full`, `@ngrx/router-store` will use the `DefaultRouterStateSerializer` serializer to serialize the Angular router event.
+When this property is set to `RouterState.Full`, `@ngrx/router-store` will use the `FullRouterStateSerializer` serializer to serialize the Angular router event.
 
 The metadata on the action will contain the Angular router event, e.g. `NavigationStart` and `RoutesRecognized`.
 
@@ -122,14 +123,9 @@ StoreRouterConnectingModule.forRoot({
 });
 </code-example>
 
-### RouterState.Minimal
+<div class="alert is-important">
 
-`RouterState.Minimal` will use the `MinimalRouterStateSerializer` serializer to serialize the Angular Router's `RouterState` and `RouterEvent`.
+The `FullRouterStateSerializer` cannot be used when [serializability runtime checks](guide/store/configuration/runtime-checks) are enabled.
+With serializability runtime checks enabled, the `DefaultRouterStateSerializer` serializer **must** be used. This also applies to Ivy with immutability runtime checks.
 
-The difference between `MinimalRouterStateSerializer` and the default `RouterStateSerializer` is that this serializer is fully serializable. To make the state and the actions serializable, the properties `paramMap`, `queryParamMap` and `component` are ignored.
-
-<code-example header="app.module.ts">
-StoreRouterConnectingModule.forRoot({
-  routerState: RouterState.Minimal,
-});
-</code-example>
+</div>
