+++ projects/ngrx.io/content/guide/router-store/index.md
@@ -21,7 +21,7 @@ import { AppComponent } from './app.component';
     RouterModule.forRoot([
       // routes
     ]),
-    // Connects RouterModule with StoreModule, uses MinimalRouterStateSerializer by default
+    // Connects RouterModule with StoreModule, uses DefaultRouterStateSerializer by default
     StoreRouterConnectingModule.forRoot(),
   ],
   bootstrap: [AppComponent],
