+++ projects/ngrx.io/content/guide/store/configuration/runtime-checks.md
@@ -39,7 +39,7 @@ export class AppModule {}
 
 <div class="alert is-important">
 
-The serializability runtime checks cannot be enabled if you use `@ngrx/router-store` with the `DefaultRouterStateSerializer`. The [default serializer](guide/router-store/configuration) has an unserializable router state and actions that are not serializable. To use the serializability runtime checks either use the `MinimalRouterStateSerializer` or implement a custom router state serializer.
+The serializability runtime checks cannot be enabled if you use `@ngrx/router-store` with the `FullRouterStateSerializer`. The [full serializer](guide/router-store/configuration) has an unserializable router state and actions that are not serializable. To use the serializability runtime checks either use the `DefaultRouterStateSerializer` or implement a custom router state serializer.
 This also applies to Ivy with immutability runtime checks.
 
 </div>
